+++ typescript/aleo-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hyperlane-xyz/aleo-sdk
 
+## 25.0.0
+
+### Patch Changes
+
+- Updated dependencies [52ce778]
+  - @hyperlane-xyz/utils@25.0.0
+  - @hyperlane-xyz/provider-sdk@1.3.1
+
 ## 24.0.0
 
 ### Patch Changes