+++ typescript/deploy-sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @hyperlane-xyz/deploy-sdk
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [52ce778]
+  - @hyperlane-xyz/utils@25.0.0
+  - @hyperlane-xyz/cosmos-sdk@25.0.0
+  - @hyperlane-xyz/aleo-sdk@25.0.0
+  - @hyperlane-xyz/provider-sdk@1.3.1
+  - @hyperlane-xyz/radix-sdk@25.0.0
+
 ## 1.3.0
 
 ### Patch Changes